@@ -19,6 +19,22 @@
 
 # Learn more: http://github.com/javan/whenever
 
-every 1.hour, :roles => [:whenever] do
+every 1.day, at: ["0:00"] :roles => [:whenever] do
   rake "podio_task:sync"
+end
+
+every 1.day, at: ["1:00"], :roles => [:whenever] do
+  rake "sync_users:sync_users"
+end
+
+every 1.day, at: ["1:15"], :roles => [:whenever] do
+  rake "sync_users:sync_plans"
+end
+
+every 1.day, at: ["1:30"], :roles => [:whenever] do
+  rake "sync_users:sync_all_progress"
+end
+
+every 1.hour, :roles => [:whenever] do
+  rake "sync_users:sync_recent_progress"
 end
