@@ -1,7 +1,11 @@
 class ApplicationController < ActionController::API
 
   def authenticate!
-    raise Pundit::NotAuthorizedError, "Not authenticated!" unless request.headers["x-authentication-token"] == Rails.application.credentials.dig(:token)
+    raise Pundit::NotAuthorizedError, "Not authenticated!" unless authenticated?
+  end
+
+  def authenticated?
+    request.headers["x-authentication-token"] == Rails.application.credentials.dig(:token) || params[:token] == Rails.application.credentials.dig(:public_token)
   end
 
   rescue_from Pundit::NotAuthorizedError do |exception|
