@@ -1,4 +1,4 @@
-class Member
+class Member < MongoidRecord
   include Mongoid::Document
   include Mongoid::Attributes::Dynamic
 
